Fix leftover top directory when no files are left
Addresses the bug where a directory is left over even though beets
should clean it up. For example, take the following structure:

```
- ImportDir
  - Artist
    - Album
      - 01 song.mp3
      - 02 song.mp3
      - digital.pdf
```

where .pdf is specified as an artifact

After performing a move the current result is:

```
- ImportDir
  - Artist
```

This is due to the fact that the plugin is invoked after beets performs
its internal cleanup and that the plugin makes its own invokation to prune_dirs, but does so
without adding the top level directory.

To remedy this, the top level directory could be added to the prune_dirs
command, but perhaps more ideal is to have the plugin execute before the
finalize call in beets as in the following snippet:

```
def manipulate_files(session, task):
    """A coroutine (pipeline stage) that performs necessary file
    manipulations *after* items have been added to the library and
    finalizes each task.
    """
    if not task.skip:
        if task.should_remove_duplicates:
            task.remove_duplicates(session.lib)

        task.manipulate_files(
            move=session.config['move'],
            copy=session.config['copy'],
            write=session.config['write'],
            link=session.config['link'],
            session=session,
        )  #### <------ MOVE PLUGIN EXECUTION TO HERE

    # Progress, cleanup, and event.
    task.finalize(session) ### <---- PLUGIN EXECUTED AFTER

```

Signed-off-by: Andrew DeMaria <lostonamountain@gmail.com>